requirements.txt
discord.py==2.3.2
openai==1.51.2
python-dotenv==1.0.0
schedule==1.2.0
aiofiles==23.2.0
aiohttp==3.9.0
